// ui/lib/api/schemas/pipelineProgress.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 */
import type { PipelineStatus } from './pipelineStatus'
import type { PipelineStep } from './pipelineStep'

export interface PipelineProgress {
  /** @minimum 0 */
  currentPage: number
  /** @minimum 0 */
  currentStepIndex: number
  jobId: string
  /** @minimum 0 */
  overallPercent: number
  status: PipelineStatus
  step?: null | PipelineStep
  /** @minimum 0 */
  totalPages: number
  /** @minimum 0 */
  totalSteps: number
}